+++ src/main/java/bridge/domain/BridgeRandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.domain;
 
 import camp.nextstep.edu.missionutils.Randoms;
