+++ src/main/java/bridge/view/ViewMessage.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.view;
 
 public enum ViewMessage {
     START_GAME_MESSAGE("다리 건너기 게임을 시작합니다."),
